# Conflicts:
#	API.postman_collection.json

+++ backend/src/main/java/es/webapp13/porthub/api/template/TemplateRestController.java
@@ -1,15 +1,19 @@
 package es.webapp13.porthub.api.template;
 
 import es.webapp13.porthub.model.Template;
+import es.webapp13.porthub.model.User;
 import es.webapp13.porthub.service.TemplateService;
+import es.webapp13.porthub.service.UserService;
 import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.net.URI;
+import java.security.Principal;
 import java.util.Optional;
 
 import static org.springframework.web.servlet.support.ServletUriComponentsBuilder.fromCurrentRequest;
@@ -21,10 +25,12 @@ public class TemplateRestController {
 
     private final TemplateService templateService;
     private final ModelMapper modelMapper;
+    private final UserService userService;
 
-    public TemplateRestController(TemplateService templateService, ModelMapper modelMapper) {
+    public TemplateRestController(TemplateService templateService, ModelMapper modelMapper, UserService userService) {
         this.templateService = templateService;
         this.modelMapper = modelMapper;
+        this.userService = userService;
     }
 
     @GetMapping("/")
@@ -46,26 +52,38 @@ public class TemplateRestController {
     }
 
     @PostMapping("/")
-    public ResponseEntity<Template> postTemplate (@RequestBody TemplateDTO templateDTO) throws IOException {
-        Template template = modelMapper.map(templateDTO, Template.class);
-        templateService.create(template);
+    public ResponseEntity<Template> postTemplate (@RequestBody TemplateDTO templateDTO, HttpServletRequest request) throws IOException {
+        Principal principal = request.getUserPrincipal();
+        Optional<User> me = userService.findById(principal.getName());
 
-        URI location = fromCurrentRequest().path("/{id}").buildAndExpand(template.getId()).toUri();
+        if (me.isPresent() && me.get().getRoles().contains("ADMIN")){
+            Template template = modelMapper.map(templateDTO, Template.class);
+            templateService.create(template);
 
-        return ResponseEntity.created(location).body(template);
+            URI location = fromCurrentRequest().path("/{id}").buildAndExpand(template.getId()).toUri();
+
+            return ResponseEntity.created(location).body(template);
+        }
+        return ResponseEntity.status(403).build();
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Template> deleteTemplate (@PathVariable long id){
-        Optional<Template> optionalTemplate = templateService.findById(id);
+    public ResponseEntity<Template> deleteTemplate (@PathVariable long id, HttpServletRequest request){
+        Principal principal = request.getUserPrincipal();
+        Optional<User> me = userService.findById(principal.getName());
 
-        if (optionalTemplate.isPresent()){
-            Template template = optionalTemplate.get();
-            templateService.delete(template);
-            return ResponseEntity.ok(template);
+        if (me.isPresent() && me.get().getRoles().contains("ADMIN")){
+            Optional<Template> optionalTemplate = templateService.findById(id);
+
+            if (optionalTemplate.isPresent()){
+                Template template = optionalTemplate.get();
+                templateService.delete(template);
+                return ResponseEntity.ok(template);
+            }
+            else {
+                return ResponseEntity.notFound().build();
+            }
         }
-        else {
-            return ResponseEntity.notFound().build();
-        }
+        return ResponseEntity.status(403).build();
     }
 }
